// packages/kindling/src/generated/ObservationInput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ObservationKind } from "./ObservationKind.js";
import type { ScopeIds } from "./ScopeIds.js";
import type { JsonValue } from "./serde_json/JsonValue.js";

/**
 * Input for creating a new observation. Optional fields are auto-generated.
 */
export type ObservationInput = { id?: string, kind: ObservationKind, content: string, provenance?: { [key in string]?: JsonValue }, ts?: number, scopeIds: ScopeIds, redacted?: boolean, };